refac and add readme
