+++ source/core/tests/unit_tests/gold/infrastructure/repositories/test_measurements_gold_repository.py
