+++ source/core/tests/unit_tests/gold/test_entry_points.py
